# frontend/openapi/identity-service.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API APSAS identity-service",
    "description": "Tài liệu API cho APSAS identity-service",
    "version": "1.0"
  },
  "servers": [{ "url": "/", "description": "mặc định" }],
...
  "paths": {
    "/api/v1/users": {
      "get": {
        "tags": ["Quản lý người dùng"],
        "summary": "Lấy tất cả người dùng",
        "description": "Lấy danh sách người dùng có phân trang và sắp xếp (chỉ Admin)",
        "operationId": "getAllUsers",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Số trang (bắt đầu từ 0)",
            "required": false,
            "schema": { "type": "string", "minimum": 0 },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số lượng phần tử mỗi trang",
            "required": false,
            "schema": { "type": "string", "maximum": 100, "minimum": 1 },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PageResponseUserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      },
      "post": {
        "tags": ["Quản lý người dùng"],
        "summary": "Tạo người dùng",
        "description": "Tạo mới người dùng (chỉ Admin)",
        "operationId": "createUser",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateUserRequest" } } },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
    "/api/v1/users/{userId}/deactivate": {
      "post": {
        "tags": ["Quản lý người dùng"],
        "summary": "Vô hiệu hóa người dùng",
        "description": "Vô hiệu hóa tài khoản người dùng (chỉ Admin)",
        "operationId": "deactivateUser",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
    "/api/v1/users/{userId}/activate": {
      "post": {
        "tags": ["Quản lý người dùng"],
        "summary": "Kích hoạt người dùng",
        "description": "Kích hoạt tài khoản người dùng (chỉ Admin)",
        "operationId": "activateUser",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
    "/api/v1/users/me/change-password": {
      "post": {
        "tags": ["Quản lý người dùng"],
        "summary": "Đổi mật khẩu",
        "description": "Đổi mật khẩu cho người dùng đang đăng nhập",
        "operationId": "changePassword",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ChangePasswordRequest" } } },
          "required": true
        },
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
...
    "/api/v1/users/me": {
      "get": {
        "tags": ["Quản lý người dùng"],
        "summary": "Lấy hồ sơ người dùng hiện tại",
        "description": "Lấy thông tin hồ sơ của người dùng đang đăng nhập",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      },
      "patch": {
        "tags": ["Quản lý người dùng"],
        "summary": "Cập nhật hồ sơ người dùng",
        "description": "Cập nhật thông tin hồ sơ của người dùng đang đăng nhập",
        "operationId": "updateCurrentUserProfile",
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateProfileRequest" } } },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "tags": ["Quản lý người dùng"],
        "summary": "Lấy người dùng theo ID",
        "description": "Lấy thông tin người dùng theo ID (chỉ Admin và Giảng viên)",
        "operationId": "getUserById",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      },
      "delete": {
        "tags": ["Quản lý người dùng"],
        "summary": "Xóa người dùng",
        "description": "Xóa tài khoản người dùng (chỉ Admin)",
        "operationId": "deleteUser",
        "parameters": [
          { "name": "userId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    },
    "/api/v1/users/role/{role}": {
      "get": {
        "tags": ["Quản lý người dùng"],
        "summary": "Lấy người dùng theo vai trò",
        "description": "Lấy danh sách người dùng theo vai trò có phân trang và sắp xếp (chỉ Admin và Giảng viên)",
        "operationId": "getUsersByRole",
        "parameters": [
          {
            "name": "role",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "enum": ["STUDENT", "INSTRUCTOR", "CONTENT_PROVIDER", "ADMIN"] }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Số trang (bắt đầu từ 0)",
            "required": false,
            "schema": { "type": "string", "minimum": 0 },
            "example": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Số lượng phần tử mỗi trang",
            "required": false,
            "schema": { "type": "string", "maximum": 100, "minimum": 1 },
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PageResponseUserResponse" } } }
          },
          "default": {
            "description": "Phản hồi lỗi mặc định",
            "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } }
          }
        },
        "security": [{ "Bearer Authentication": [] }]
      }
    }
  },
  "components": {
    "schemas": {
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email", "minLength": 1 },
          "password": { "type": "string", "maxLength": 2147483647, "minLength": 8 },
          "firstName": { "type": "string", "maxLength": 100, "minLength": 0 },
          "lastName": { "type": "string", "maxLength": 100, "minLength": 0 },
          "role": { "type": "string", "enum": ["STUDENT", "INSTRUCTOR", "CONTENT_PROVIDER", "ADMIN"] },
          "isActive": { "type": "boolean" },
          "isEmailVerified": { "type": "boolean" }
        },
        "required": ["email", "firstName", "lastName", "password"]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "role": { "type": "string", "enum": ["STUDENT", "INSTRUCTOR", "CONTENT_PROVIDER", "ADMIN"] },
          "isActive": { "type": "boolean" },
          "isEmailVerified": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "currentPassword": { "type": "string", "minLength": 1 },
          "newPassword": { "type": "string", "maxLength": 2147483647, "minLength": 8 }
        },
        "required": ["currentPassword", "newPassword"]
      },
...
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string", "maxLength": 100, "minLength": 0 },
          "lastName": { "type": "string", "maxLength": 100, "minLength": 0 }
        }
      },
      "PageResponseUserResponse": {
        "type": "object",
        "description": "Lớp bọc phản hồi phân trang",
        "properties": {
          "content": {
            "type": "array",
            "description": "Danh sách phần tử trong trang hiện tại",
            "items": { "$ref": "#/components/schemas/UserResponse" }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32",
            "description": "Số trang hiện tại (bắt đầu từ 0)",
            "example": 0
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "description": "Số lượng phần tử mỗi trang",
            "example": 10
          },
          "totalElements": {
            "type": "integer",
            "format": "int64",
            "description": "Tổng số phần tử trên tất cả các trang",
            "example": 100
          },
          "totalPages": { "type": "integer", "format": "int32", "description": "Tổng số trang", "example": 10 },
          "first": { "type": "boolean", "description": "Có phải trang đầu tiên không", "example": true },
          "last": { "type": "boolean", "description": "Có phải trang cuối cùng không", "example": false },
          "hasNext": { "type": "boolean", "description": "Có trang tiếp theo không", "example": true },
          "hasPrevious": { "type": "boolean", "description": "Có trang trước đó không", "example": false }
        }
      },
      "ProblemDetail": {
        "description": "Chi tiết lỗi theo chuẩn RFC 9457",
        "properties": {
          "type": { "format": "uri", "description": "Loại lỗi (URI)" },
          "title": { "description": "Tiêu đề lỗi" },
          "status": { "format": "int32", "description": "Mã trạng thái HTTP" },
          "detail": { "description": "Mô tả chi tiết lỗi" },
          "instance": { "format": "uri", "description": "Đường dẫn instance lỗi" }
        }
      }
    },
    "securitySchemes": { "Bearer Authentication": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" } }
  }
}